@@ -1,40 +1,41 @@
 name: unicaen_timetable
 description: Allows the students of the University of Caen Normandie to check their timetable directly from their smartphone.
-version: 3.1.0+78
+version: 3.2.0-beta.1+79
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^4.3.2
-  flutter_week_view: ^1.0.0
-  flutter_svg: ^0.19.1
-  hive: ^1.4.4
-  hive_flutter: ^0.3.1
-  ez_localization: ^0.3.0
-  flutter_secure_storage: ^3.3.5
-  steel_crypt: ^2.2.2
-  path_provider: ^1.6.21
-  pedantic: ^1.9.2
-  http: ^0.12.2
-  background_fetch: ^0.6.0
-  flutter_colorpicker: ^0.3.4
-  implicitly_animated_reorderable_list: ^0.3.0
-  share: ^0.6.5
-  rate_my_app: ^0.7.2
-  url_launcher: ^5.4.11
-  package_info: ^0.4.3
-  device_info: ^0.4.2
-  scrollable_positioned_list: ^0.1.8
-  admob_flutter: ^1.0.0
-  catcher: ^0.3.20
+  provider: ^5.0.0
+  flutter_week_view: ^1.1.0
+  flutter_svg: ^0.21.0-nullsafety.0
+  hive: ^2.0.0
+  hive_flutter: ^1.0.0
+  ez_localization: ^1.0.0
+  flutter_secure_storage: ^4.1.0
+  encrypt: ^5.0.0-beta.1
+  path_provider: ^2.0.1
+  pedantic: ^1.11.0
+  http: ^0.13.1
+  background_fetch: ^1.0.0-nullsafety.1
+  flutter_colorpicker: ^0.4.0
+  implicitly_animated_reorderable_list: ^0.4.0-dev+4
+  share: ^2.0.1
+  rate_my_app: ^1.1.0
+  url_launcher: ^6.0.2
+  package_info: ^2.0.0
+  device_info: ^2.0.0
+  scrollable_positioned_list: ^0.2.0-nullsafety.0
+  admob_flutter: ^2.0.0-nullsafety.0
+  catcher: ^0.6.3
+  collection: ^1.15.0
 
 dev_dependencies:
-  hive_generator: ^0.8.1
-  build_runner: ^1.10.3
+  hive_generator: ^1.0.0
+  build_runner: ^1.11.5
 
 flutter:
   uses-material-design: true
