@@ -9,7 +9,7 @@ import 'package:unicaen_timetable/model/settings/entries/entry.dart';
 class LessonNotificationModeSettingsEntry extends SettingsEntry<int> {
   /// Creates a new app notification mode settings entry instance.
   LessonNotificationModeSettingsEntry({
-    @required String keyPrefix,
+    required String keyPrefix,
   }) : super(
           keyPrefix: keyPrefix,
           key: 'lesson_notification_mode',
@@ -28,27 +28,31 @@ class _LessonNotificationModeSettingsEntryWidget extends StatelessWidget {
 
   /// Creates a new lesson notification mode settings entry widget instance.
   const _LessonNotificationModeSettingsEntryWidget({
-    @required this.entry,
+    required this.entry,
   });
 
   @override
   Widget build(BuildContext context) => SettingsDropdownButton<int>(
         titleKey: 'settings.${entry.key}',
         onChanged: (value) async {
-          bool result = await UnicaenTimetableApp.CHANNEL.invokeMethod('activity.lesson_notification_mode_changed', {'value': value});
-          if (result) {
+          if (value == null) {
+            return;
+          }
+
+          bool? result = await UnicaenTimetableApp.CHANNEL.invokeMethod<bool>('activity.lesson_notification_mode_changed', {'value': value});
+          if (result != null && result) {
             entry.value = value;
             await entry.flush();
           }
         },
         items: [
           DropdownMenuItem<int>(
-            child: Text(context.getString('other.lesson_notification_mode.disabled')),
             value: -1,
+            child: Text(context.getString('other.lesson_notification_mode.disabled')),
           ),
           DropdownMenuItem<int>(
-            child: Text(context.getString('other.lesson_notification_mode.alarms_only')),
             value: 0,
+            child: Text(context.getString('other.lesson_notification_mode.alarms_only')),
           ),
         ],
         value: entry.value,
