@@ -24,15 +24,15 @@ class SettingsEntry<T> extends ChangeNotifier with RenderableSettingsObject {
   /// Creates a new settings entry instance.
   SettingsEntry({
     String keyPrefix = '',
-    @required String key,
+    required String key,
     this.mutable = true,
     this.enabled = true,
-    @required T value,
+    required T value,
   })  : key = (keyPrefix.isEmpty ? '' : (keyPrefix + '.')) + key,
         _value = value;
 
   /// Loads this entry value from the settings box.
-  Future<void> load([Box settingsBox]) async {
+  Future<void> load([Box? settingsBox]) async {
     Box box = settingsBox ?? await Hive.openBox(SettingsModel.HIVE_BOX);
     _value = decodeValue(box.get(key));
   }
@@ -51,7 +51,7 @@ class SettingsEntry<T> extends ChangeNotifier with RenderableSettingsObject {
   }
 
   /// Flushes this entry value to the settings box.
-  Future<void> flush([Box settingsBox]) async {
+  Future<void> flush([Box? settingsBox]) async {
     if (!mutable) {
       return;
     }
@@ -75,7 +75,7 @@ class SettingsEntryWidget extends StatelessWidget {
 
   /// Creates a new settings entry widget instance.
   const SettingsEntryWidget({
-    @required this.entry,
+    required this.entry,
   });
 
   @override
@@ -87,7 +87,7 @@ class SettingsEntryWidget extends StatelessWidget {
       );
 
   /// Creates the subtitle widget.
-  Widget createSubtitle(BuildContext context) {
+  Widget? createSubtitle(BuildContext context) {
     if (entry.value is String) {
       return Text(entry.value == null || entry.value.isEmpty ? context.getString('other.empty') : entry.value);
     }
@@ -95,7 +95,7 @@ class SettingsEntryWidget extends StatelessWidget {
   }
 
   /// Creates the controller widget.
-  Widget createController(BuildContext context) {
+  Widget? createController(BuildContext context) {
     if (entry.value is bool) {
       return Switch(
         value: entry.value,
@@ -122,7 +122,7 @@ class SettingsEntryWidget extends StatelessWidget {
     }
 
     if (entry.value is String) {
-      String value = await TextInputDialog.getValue(
+      String? value = await TextInputDialog.getValue(
         context,
         titleKey: 'settings.${entry.key}',
         initialValue: entry.value,
@@ -159,14 +159,14 @@ class SettingsDropdownButton<T> extends StatelessWidget {
   final T value;
 
   /// Triggered when the value has been changed.
-  final Function(T value) onChanged;
+  final ValueChanged<T?>? onChanged;
 
   /// Creates a new settings dropdown button instance.
   const SettingsDropdownButton({
-    @required this.titleKey,
-    @required this.items,
-    @required this.value,
-    @required this.onChanged,
+    required this.titleKey,
+    required this.items,
+    required this.value,
+    this.onChanged,
   });
 
   @override
